devtools/client/shared/components/reps/test/mochitest/test_reps_attribute.html

<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<!DOCTYPE HTML>
<html>
<!--
Test Attribute rep
-->
<head>
  <meta charset="utf-8">
  <title>Rep test - Attribute</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="head.js" type="application/javascript;version=1.8"></script>
<script type="application/javascript;version=1.8">
window.onload = Task.async(function* () {
  try {
    let { Rep } = browserRequire("devtools/client/shared/components/reps/rep");
    let { Attribute } = browserRequire("devtools/client/shared/components/reps/attribute");

    let gripStub = {
      "type": "object",
      "class": "Attr",
      "actor": "server1.conn19.obj65",
      "extensible": true,
      "frozen": false,
      "sealed": false,
      "ownPropertyLength": 0,
      "preview": {
        "kind": "DOMNode",
        "nodeType": 2,
        "nodeName": "class",
        "value": "autocomplete-suggestions"
      }
    };

    // Test that correct rep is chosen
    const renderedRep = shallowRenderComponent(Rep, { object: gripStub });
    is(renderedRep.type, Attribute.rep, `Rep correctly selects ${Attribute.rep.displayName}`);

    // Test rendering
    const renderedComponent = renderComponent(Attribute.rep, { object: gripStub });
    is(renderedComponent.textContent, "class=\"autocomplete-suggestions\"", "Attribute rep has expected text content");
  } catch(e) {
    ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
  } finally {
    SimpleTest.finish();
  }
});
</script>
</pre>
</body>
</html>
